refactoring ImageClassSet out of ImageNet
